pyproject.toml
[project]
name = "microblog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosmtpd>=1.4.6",
    "bs4>=0.0.2",
    "email-validator>=2.3.0",
    "feedgen>=1.0.0",
    "flask>=3.1.2",
    "flask-httpauth>=4.8.0",
    "flask-login>=0.6.3",
    "flask-mail>=0.10.0",
    "flask-migrate>=4.1.0",
    "flask-moment>=1.0.6",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "httpie>=3.2.4",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
